Replace nc[mpi]_delete calls
Replacing nc[mpi]_delete calls with POSIX functions to delete
files and directories as needed.

We have been using these functions when ADIOS support is enabled,
but now use it (instead of [P]netCDF functions) for all cases.

Also changing the barrier from just on I/O processes to all
processes (Since potentially after a delete another comm might
expect the file being deleted), before/after a delete.

+++ src/clib/pio_file.cpp
@@ -6,6 +6,8 @@
 #include <pio.h>
 #include <pio_internal.h>
 #include "spio_io_summary.h"
+#include <sys/stat.h>
+#include <unistd.h>
 
 #ifdef _ADIOS2
 #include "../../tools/adios2pio-nm/adios2pio-nm-lib-c.h"
@@ -1155,12 +1157,12 @@ int PIOc_deletefile_impl(int iosysid, const char *filename)
      * barriers are needed to assure that no task is trying to operate
      * on the file while it is being deleted. IOTYPE is not known, but
      * nc_delete() will delete any type of file. */
+    mpierr = MPI_Barrier(ios->union_comm);
     if (ios->ioproc)
     {
-        mpierr = MPI_Barrier(ios->io_comm);
-
         if (!mpierr && ios->io_rank == 0)
         {
+            struct stat sd;
 #ifdef _ADIOS2
             /* Append ".bp" to filename for the corresponding ADIOS BP filename */
             static const char adios_bp_filename_extn[] = ".bp";
@@ -1175,27 +1177,27 @@ int PIOc_deletefile_impl(int iosysid, const char *filename)
             }
             snprintf(adios_bp_filename, adios_bp_filename_len, "%s%s", filename, adios_bp_filename_extn);
 
-            struct stat sd;
             if (0 == stat(adios_bp_filename, &sd))
             {
                 spio_remove_directory(adios_bp_filename);
             }
             free(adios_bp_filename);
-
-            /* Delete the file (for ADIOS BP files, delete the symlink file) */
-            ierr = unlink(filename);
-#elif defined(_PNETCDF)
-            ierr = ncmpi_delete(filename, MPI_INFO_NULL);
-#elif defined(_NETCDF)
-            ierr = nc_delete(filename);
-#else
-            ierr = unlink(filename);
 #endif
+
+            if(stat(filename, &sd) == 0){
+              if(S_ISDIR(sd.st_mode)){
+                /* Delete the directory pointed to by filename (e.g. NCZarr files) */
+                spio_remove_directory(filename);
+              }
+              else{
+                /* Delete the file (for ADIOS BP files, delete the symlink file) */
+                ierr = unlink(filename);
+              }
+            }
         }
 
-        if (!mpierr)
-            mpierr = MPI_Barrier(ios->io_comm);
     }
+    mpierr = MPI_Barrier(ios->union_comm);
     LOG((2, "PIOc_deletefile ierr = %d", ierr));
 
     ierr = check_netcdf(ios, NULL, ierr, __FILE__, __LINE__);
